requirements.txt
fastapi[standard]==0.112.0
uvicorn==0.22.0
requests==2.32.3
python-dotenv==1.0.0